fix: add conventional-changelog preset for pnpm
The conventional-changelog presets are resolved by a preset loader inside the conventional-changelog
tooling. Based on yarn and npm installing them inside the node_modules folder this works fine for
them. But pnpm only symlinks what is really defined inside the package.json and does not rely on
implicit package installation.